refactor: Name the layer ids the GBTS connector reader parses
The connector file reader read `src` and `dst` into bare `std::uint32_t`
and handed them to `GbtsLayerConnection`, whose members are
`GbtsExperimentLayerId`; the unit test cast ids that already carry the
type. Both now spell the type the rest of the pass introduced.

+++ Tests/UnitTests/Core/Seeding/GbtsSeedingTests.cpp
@@ -153,8 +153,7 @@ std::shared_ptr<Experimental::GbtsGeometry> makeGeometry(
   std::vector<Experimental::GbtsLayerConnection> connections;
   connections.reserve(detector.links.size());
   for (const auto& [src, dst] : detector.links) {
-    connections.push_back(
-        {static_cast<std::uint32_t>(src), static_cast<std::uint32_t>(dst)});
+    connections.push_back({src, dst});
   }
 
   return std::make_shared<Experimental::GbtsGeometry>(layers, connections,
